+++ pages/index.js
@@ -1,32 +1,55 @@
-// import fs from "fs";
-// import matter from "gray-matter";
-// import path from "path";
+import fs from "fs";
+import matter from "gray-matter";
+import path from "path";
 import Post from "@components/Post";
-import supabase from "@utils/initSupabase";
+import { sortByDate } from "@utils/index";
 import Container from "@components/Container";
 
-export default function Home({ blogPosts }) {
+const Home = ({ posts }) => {
   return (
     <>
       <Container>
         <div>
-          {blogPosts.map((blogPosts) => (
-            <Post post={blogPosts} key={blogPosts.slug} />
+          {posts.map((post, index) => (
+            <Post post={post} key={index} />
           ))}
         </div>
       </Container>
     </>
   );
-}
+};
 
-export async function getServerSideProps() {
-  const { data } = await supabase
-    .from("blogPosts")
-    .select("id, title, date, image, description, body, slug");
+// The component Post is able to accept the props generated by the following getStaticprops
+// It maps through the array produced by getStaticprops
+
+export async function getStaticProps() {
+  // get files from the posts dir
+  const files = fs.readdirSync(path.join("posts"));
+
+  // Get slug and frontmatter from posts
+  const posts = files.map((filename) => {
+    //create slug
+
+    const slug = filename.replace(".md", "");
+
+    // Get frontmatter
+    const markdownWithMeta = fs.readFileSync(
+      path.join("posts", filename),
+      "utf-8"
+    );
+    const { data: frontmatter } = matter(markdownWithMeta);
+
+    return {
+      slug,
+      frontmatter,
+    };
+  });
 
   return {
     props: {
-      blogPosts: data,
+      posts: posts.sort(sortByDate),
     },
   };
 }
+
+export default Home;
